Refactor tutorial workflow helpers
