+++ docs/source/_static/examples/python/tutorial_choose_active_space.py
@@ -14,16 +14,17 @@ CASCI energy.
 # --------------------------------------------------------------------------------------------
 
 from dataclasses import dataclass
-from functools import partial
-from itertools import combinations
 from typing import cast
 
-import numpy as np
 from qdk_chemistry.algorithms import create
-from qdk_chemistry.data import MajoranaMapping, Orbitals, Structure, Wavefunction
+from qdk_chemistry.data import Orbitals, Structure, Wavefunction
 from qdk_chemistry.data.symmetry import SymmetryLabel, axes
 from qdk_chemistry.utils import Logger, compute_valence_space_parameters
 from qdk_chemistry.utils.cubegen import generate_cubefiles_from_orbitals
+from tutorial_orbital_coordinates import (
+    NaturalOrbitalCoordinateMinimizationResult,
+    coordinate_minimize_natural_orbital_coefficient_norm,
+)
 
 
 @dataclass
@@ -86,437 +87,6 @@ class ActiveSpaceResult:
     num_refined_determinants: int
     natural_orbital_coordinate_minimization: (
         "NaturalOrbitalCoordinateMinimizationResult"
-    )
-
-
-@dataclass(frozen=True)
-class NaturalOrbitalCoordinateMinimizationResult:
-    """Orbitals and diagnostics from deterministic coordinate minimization.
-
-    Here *gauge* means a choice of orbital vectors used as coordinates for a
-    fixed occupation-degenerate subspace. Different gauges span the same
-    subspace and have the same exact CASCI energy, but can produce different
-    mapped Hamiltonian representations.
-
-    Attributes:
-        orbitals: Selected orbitals after AO anchoring and coordinate minimization.
-        selected_blocks: Natural-orbital index blocks present in the selected
-            active space. Singleton blocks require no rotational search.
-        coefficient_norm_before: Mapped coefficient norm ``lambda`` in Hartree
-            after AO anchoring and before coordinate rotations.
-        coefficient_norm_after: Mapped coefficient norm in Hartree after accepted
-            coordinate rotations.
-        effective_pauli_terms_before: Pauli terms above the diagnostic threshold
-            after AO anchoring and before coordinate rotations.
-        effective_pauli_terms_after: Corresponding effective term count after
-            gauge selection; this is reported but not optimized directly.
-    """
-
-    orbitals: Orbitals
-    selected_blocks: tuple[tuple[int, ...], ...]
-    coefficient_norm_before: float
-    coefficient_norm_after: float
-    effective_pauli_terms_before: int
-    effective_pauli_terms_after: int
-
-
-def _natural_occupation_blocks(
-    reference_wavefunction: Wavefunction,
-    natural_indices: list[int],
-    *,
-    degeneracy_tolerance: float = 1e-6,
-) -> tuple[tuple[int, ...], ...]:
-    """Group consecutive sorted natural occupations into degeneracy blocks.
-
-    Natural occupations are eigenvalues of the spin-traced active-space
-    one-particle RDM and are returned in descending order. Consecutive values
-    separated by less than ``degeneracy_tolerance`` define one block whose
-    eigenvectors can rotate without changing the RDM.
-
-    Args:
-        reference_wavefunction: Correlated wavefunction containing the active
-            spin-traced one-particle RDM.
-        natural_indices: Orbital indices ordered like the descending natural
-            occupations.
-        degeneracy_tolerance: Maximum dimensionless occupation-number gap within
-            one block.
-
-    Returns:
-        Non-overlapping ordered blocks that cover ``natural_indices``.
-    """
-    one_rdm = np.asarray(reference_wavefunction.get_active_one_rdm_spin_traced())
-    occupations = np.linalg.eigvalsh(0.5 * (one_rdm + one_rdm.T))[::-1]
-
-    occupation_blocks: list[tuple[int, ...]] = []
-    block_start = 0
-    for block_stop in range(1, len(occupations) + 1):
-        if block_stop == len(occupations) or (
-            abs(occupations[block_stop] - occupations[block_stop - 1])
-            >= degeneracy_tolerance
-        ):
-            occupation_blocks.append(tuple(natural_indices[block_start:block_stop]))
-            block_start = block_stop
-    return tuple(occupation_blocks)
-
-
-def _orbitals_with_coefficients(
-    template: Orbitals, coefficients: np.ndarray
-) -> Orbitals:
-    """Copy an orbital object while replacing its coefficient matrix.
-
-    Args:
-        template: Source of overlap, basis-set, and active/inactive metadata.
-        coefficients: Replacement AO-by-MO coefficient matrix.
-
-    Returns:
-        Restricted orbitals with replacement coefficients and metadata copied
-        from ``template``. Orbital energies are omitted because rotations inside
-        a correlated active space do not assign unique one-electron energies.
-    """
-    return Orbitals(
-        coefficients,
-        None,
-        template.get_overlap_matrix(),
-        template.get_basis_set(),
-        template.active_indices(),
-        template.inactive_indices(),
-    )
-
-
-def _ao_anchor_block(block: np.ndarray, overlap: np.ndarray) -> np.ndarray:
-    """Orient one natural-orbital block reproducibly using atomic-orbital anchors.
-
-    Degenerate eigenvectors can arrive in any orthonormal orientation. The AO
-    projections depend only on the subspace projector, so pivoting those
-    projections and symmetrically orthogonalizing them produces the same block
-    after arbitrary rotations or sign changes of the input eigenvectors.
-
-    Args:
-        block: AO-by-orbital coefficient matrix for one occupation-degenerate block.
-        overlap: Atomic-orbital overlap matrix.
-
-    Returns:
-        An AO-by-orbital matrix spanning the same subspace, with a deterministic
-        orientation and columns orthonormal in the AO metric.
-
-    Raises:
-        RuntimeError: If independent AO anchors cannot be found.
-    """
-    projected_ao = overlap @ block
-    residuals = projected_ao.copy()
-    anchors = []
-
-    # Pivoted Gram-Schmidt chooses independent AO rows. Rounding the projection
-    # norms makes near-ties reproducible; np.argmax then selects the lowest AO
-    # index among equal rounded values. The coefficients themselves are not rounded.
-    for _ in range(block.shape[1]):
-        projection_norms = np.einsum("ij,ij->i", residuals, residuals)
-        anchor = int(np.argmax(np.round(projection_norms, decimals=14)))
-        anchors.append(anchor)
-        anchor_vector = residuals[anchor].copy()
-        anchor_norm = np.linalg.norm(anchor_vector)
-        if anchor_norm <= np.finfo(float).eps:
-            raise RuntimeError(
-                "Unable to find independent AO anchors for a natural-orbital block"
-            )
-        anchor_vector /= anchor_norm
-        residuals -= np.outer(residuals @ anchor_vector, anchor_vector)
-
-    # Löwdin symmetric orthogonalization turns the projected AO anchors into an
-    # orthonormal rotation without privileging their processing order.
-    anchor_coefficients = projected_ao[anchors].T
-    gram_values, gram_vectors = np.linalg.eigh(
-        anchor_coefficients.T @ anchor_coefficients
-    )
-    if np.min(gram_values) <= np.finfo(float).eps:
-        raise RuntimeError("Natural-orbital AO anchors are linearly dependent")
-    orthogonalizer = (
-        anchor_coefficients @ gram_vectors @ np.diag(gram_values**-0.5) @ gram_vectors.T
-    )
-    return block @ orthogonalizer
-
-
-def _givens_rotation(size: int, left: int, right: int, angle: float) -> np.ndarray:
-    """Construct an orthogonal plane rotation inside an orbital block.
-
-    Args:
-        size: Dimension of the degenerate orbital block.
-        left: First zero-based block-column index to rotate.
-        right: Second zero-based block-column index to rotate.
-        angle: Rotation angle in radians.
-
-    Returns:
-        A ``size``-by-``size`` identity matrix with the selected plane replaced
-        by a Givens rotation.
-
-    Raises:
-        IndexError: If either selected column lies outside the matrix.
-    """
-    rotation = np.eye(size)
-    cosine = np.cos(angle)
-    sine = np.sin(angle)
-    rotation[left, left] = cosine
-    rotation[left, right] = -sine
-    rotation[right, left] = sine
-    rotation[right, right] = cosine
-    return rotation
-
-
-def _golden_section_minimum(
-    objective,
-    lower_bound: float,
-    upper_bound: float,
-    *,
-    argument_tolerance: float = 1e-13,
-) -> tuple[float, float]:
-    """Refine a bracketed scalar minimum without a machine-epsilon floor.
-
-    The mapped coefficient norm is a sum of absolute values and can therefore
-    have a cusp where symmetry-related Pauli coefficients vanish. SciPy's
-    bounded minimizer stops at an argument tolerance proportional to
-    ``sqrt(machine epsilon)``, which leaves platform-dependent residual terms
-    near such cusps. Golden-section contraction uses the requested absolute
-    angular tolerance directly.
-
-    Args:
-        objective: Scalar function to minimize.
-        lower_bound: Inclusive lower end of the bracketing interval.
-        upper_bound: Inclusive upper end of the bracketing interval.
-        argument_tolerance: Maximum final interval width.
-
-    Returns:
-        The best sampled argument and its objective value.
-
-    Raises:
-        ValueError: If the interval or tolerance is not positive.
-    """
-    if upper_bound <= lower_bound:
-        raise ValueError("upper_bound must be greater than lower_bound")
-    if argument_tolerance <= 0.0:
-        raise ValueError("argument_tolerance must be positive")
-
-    inverse_golden_ratio = (np.sqrt(5.0) - 1.0) / 2.0
-    left = float(lower_bound)
-    right = float(upper_bound)
-    inner_left = right - inverse_golden_ratio * (right - left)
-    inner_right = left + inverse_golden_ratio * (right - left)
-    value_left = float(objective(inner_left))
-    value_right = float(objective(inner_right))
-
-    while right - left > argument_tolerance:
-        if value_left <= value_right:
-            right = inner_right
-            inner_right = inner_left
-            value_right = value_left
-            inner_left = right - inverse_golden_ratio * (right - left)
-            value_left = float(objective(inner_left))
-        else:
-            left = inner_left
-            inner_left = inner_right
-            value_left = value_right
-            inner_right = left + inverse_golden_ratio * (right - left)
-            value_right = float(objective(inner_right))
-
-    candidates = (
-        (inner_left, value_left),
-        (inner_right, value_right),
-        ((left + right) / 2.0, float(objective((left + right) / 2.0))),
-    )
-    return min(candidates, key=lambda candidate: (candidate[1], candidate[0]))
-
-
-def coordinate_minimize_natural_orbital_coefficient_norm(
-    reference_wavefunction: Wavefunction,
-    selected_orbitals: Orbitals,
-    natural_indices: list[int],
-    *,
-    degeneracy_tolerance: float = 1e-6,
-    angle_samples: int = 32,
-    max_sweeps: int = 3,
-    improvement_tolerance: float = 1e-10,
-    effective_pauli_threshold: float = 1e-10,
-) -> NaturalOrbitalCoordinateMinimizationResult:
-    r"""Coordinate-minimize :math:`\lambda` inside degenerate orbital blocks.
-
-    Natural orbitals with equal occupations define a subspace but not unique
-    orbital vectors. First, AO anchoring gives every selected occupation block
-    deterministic coordinates. Then deterministic coordinate sweeps search
-    plane rotations within degenerate blocks and accept only reductions in the
-    mapped coefficient norm :math:`\lambda=\sum_\ell |h_\ell|`.
-
-    Rotations stay inside occupation-degenerate blocks, so they preserve the
-    selected orbital subspace and its exact CASCI energy. Minimizing lambda is
-    useful for phase-estimation time bounds and future Hamiltonian-simulation
-    cost analysis. Effective Pauli counts are diagnostics rather than the
-    optimization objective because raw counts are sensitive to pruning noise.
-
-    The search is deterministic coordinate descent, not a proof of the global
-    minimum for arbitrary high-dimensional blocks. Each sweep visits every
-    Givens plane in a fixed order. A coarse angular grid identifies a candidate
-    basin and golden-section contraction refines it without a machine-epsilon
-    stopping floor. Only reductions larger than ``improvement_tolerance`` are
-    accepted. For the two-dimensional pi
-    blocks in the diatomic examples, each coordinate is a one-angle search;
-    larger blocks can have coupled local minima.
-
-    Args:
-        reference_wavefunction: Correlated valence-space wavefunction whose
-            spin-traced one-particle RDM defines occupation degeneracies.
-        selected_orbitals: autoCAS-selected natural orbitals to canonicalize.
-        natural_indices: Valence orbital indices ordered like the natural
-            occupations in ``reference_wavefunction``.
-        degeneracy_tolerance: Dimensionless occupation-number gap below which
-            consecutive natural orbitals belong to one degenerate block.
-        angle_samples: Uniform samples over :math:`[0,\pi)` used to locate a
-            candidate basin for each plane rotation. Must be at least four.
-        max_sweeps: Maximum deterministic passes over all block-coordinate planes.
-        improvement_tolerance: Minimum absolute reduction in :math:`\lambda`, in
-            Hartree, required to accept a rotation or begin another sweep.
-        effective_pauli_threshold: Absolute Pauli-coefficient threshold, in
-            Hartree, used only for before/after diagnostic term counts.
-
-    Returns:
-        Coordinate-minimized orbitals and before/after mapping diagnostics.
-
-    Raises:
-        ValueError: If fewer than four angular samples are requested.
-        RuntimeError: If autoCAS splits an occupation-degenerate block or AO
-            anchoring cannot construct deterministic coordinates.
-    """
-    if angle_samples < 4:
-        raise ValueError("angle_samples must be at least four")
-
-    alpha_channel = SymmetryLabel([axes.alpha()])
-    selected_indices = tuple(selected_orbitals.active_indices().indices(alpha_channel))
-    selected_index_set = set(selected_indices)
-    selected_blocks = []
-    for block in _natural_occupation_blocks(
-        reference_wavefunction,
-        natural_indices,
-        degeneracy_tolerance=degeneracy_tolerance,
-    ):
-        selected_members = selected_index_set.intersection(block)
-        if selected_members and len(selected_members) != len(block):
-            raise RuntimeError(
-                f"The selected active space splits degenerate natural-orbital block {block}"
-            )
-        if selected_members:
-            selected_blocks.append(block)
-
-    coefficients = np.array(
-        selected_orbitals.coefficients().block((alpha_channel, alpha_channel)),
-        copy=True,
-    )
-    overlap = selected_orbitals.get_overlap_matrix()
-    # AO anchoring fixes signs, permutations, and common rotations before the
-    # lambda search. If lambda has equivalent minima, keeping a zero rotation
-    # below preserves this deterministic AO-based tie-break.
-    for block in selected_blocks:
-        coefficients[:, block] = _ao_anchor_block(coefficients[:, block], overlap)
-
-    hamiltonian_constructor = create("hamiltonian_constructor", "qdk")
-    qubit_mapper = create(
-        "qubit_mapper",
-        "qdk",
-        threshold=1e-14,
-        integral_threshold=1e-14,
-    )
-    mapping = MajoranaMapping.jordan_wigner(num_modes=2 * len(selected_indices))
-
-    def mapped_metrics(candidate_coefficients: np.ndarray) -> tuple[float, int]:
-        """Map one candidate and return ``(lambda, effective Pauli terms)``.
-
-        Mapping uses tight ``1e-14`` integral and Pauli thresholds so pruning
-        does not create artificial minima. The separately reported effective
-        count applies ``effective_pauli_threshold`` for a stable diagnostic.
-        """
-        candidate_orbitals = _orbitals_with_coefficients(
-            selected_orbitals, candidate_coefficients
-        )
-        hamiltonian = hamiltonian_constructor.run(candidate_orbitals)
-        qubit_hamiltonian = qubit_mapper.run(hamiltonian, mapping)
-        effective_terms = int(
-            np.count_nonzero(
-                np.abs(qubit_hamiltonian.coefficients) >= effective_pauli_threshold
-            )
-        )
-        return float(qubit_hamiltonian.schatten_norm), effective_terms
-
-    def rotated_norm(
-        angle: float,
-        *,
-        base_coefficients: np.ndarray,
-        block_coefficients: np.ndarray,
-        block: tuple[int, ...],
-        left: int,
-        right: int,
-    ) -> float:
-        """Return lambda after one candidate Givens rotation.
-
-        ``partial`` binds the current block, coordinate plane, and coefficient
-        snapshots, leaving a one-argument objective for scalar minimization.
-        """
-        candidate_coefficients = base_coefficients.copy()
-        candidate_coefficients[:, block] = block_coefficients @ _givens_rotation(
-            len(block), left, right, angle
-        )
-        return mapped_metrics(candidate_coefficients)[0]
-
-    norm_before, terms_before = mapped_metrics(coefficients)
-    current_norm = norm_before
-    angle_grid = np.linspace(0.0, np.pi, angle_samples, endpoint=False)
-    angle_step = np.pi / angle_samples
-
-    # Coordinate descent supports blocks larger than two through Givens plane
-    # rotations. Diatomic pi blocks are two-dimensional, so each sweep reduces
-    # to a transparent one-angle search.
-    for _ in range(max_sweeps):
-        sweep_start_norm = current_norm
-        for block in selected_blocks:
-            if len(block) < 2:
-                continue
-            for left, right in combinations(range(len(block)), 2):
-                block_coefficients = coefficients[:, block].copy()
-                norm_at_angle = partial(
-                    rotated_norm,
-                    base_coefficients=coefficients,
-                    block_coefficients=block_coefficients,
-                    block=block,
-                    left=left,
-                    right=right,
-                )
-
-                grid_norms = np.asarray(
-                    [norm_at_angle(float(angle)) for angle in angle_grid]
-                )
-                best_index = int(np.argmin(grid_norms))
-                best_angle = float(angle_grid[best_index])
-                # Refine even when the best coarse sample is the current
-                # zero-angle point. A symmetry cusp can lie within 1e-8 radians
-                # of that sample, below the grid's resolution but far enough
-                # away to leave platform-dependent Pauli coefficients.
-                refined_angle, refined_norm = _golden_section_minimum(
-                    norm_at_angle,
-                    best_angle - angle_step,
-                    best_angle + angle_step,
-                )
-                if refined_norm < current_norm - improvement_tolerance:
-                    coefficients[:, block] = block_coefficients @ _givens_rotation(
-                        len(block), left, right, float(refined_angle % np.pi)
-                    )
-                    current_norm = refined_norm
-
-        if current_norm >= sweep_start_norm - improvement_tolerance:
-            break
-
-    norm_after, terms_after = mapped_metrics(coefficients)
-    return NaturalOrbitalCoordinateMinimizationResult(
-        orbitals=_orbitals_with_coefficients(selected_orbitals, coefficients),
-        selected_blocks=tuple(selected_blocks),
-        coefficient_norm_before=norm_before,
-        coefficient_norm_after=norm_after,
-        effective_pauli_terms_before=terms_before,
-        effective_pauli_terms_after=terms_after,
     )
 
 
@@ -649,6 +219,12 @@ N    0.000000    0.000000    1.850000
     # Summarize the inactive, selected active, and virtual spatial-orbital spaces.
     alpha_channel = SymmetryLabel([axes.alpha()])
     refined_indices = list(refined_orbitals.active_indices().indices(alpha_channel))
+    if not refined_indices:
+        raise RuntimeError(
+            "autoCAS selected no active orbitals. Set refined_wavefunction to "
+            "natural_orbital_wavefunction to retain the complete valence space, "
+            "or adjust the autoCAS thresholds before continuing."
+        )
     inactive_indices = list(refined_orbitals.inactive_indices().indices(alpha_channel))
     num_refined_alpha, num_refined_beta = (
         refined_wavefunction.get_active_num_electrons()
