+++ docs/source/_static/examples/python/tutorial_run_iqpe.py
@@ -163,7 +163,8 @@ def choose_reference_guided_evolution_time(
         Every intermediate and final quantity needed to inspect the derivation.
 
     Raises:
-        ValueError: If ``num_phase_bits`` is not positive.
+        ValueError: If the precision, Hamiltonian norm, selected phase-grid
+            point, target energy, or resulting evolution time is unusable.
     """
     if num_phase_bits <= 0:
         raise ValueError("num_phase_bits must be positive")
@@ -171,12 +172,27 @@ def choose_reference_guided_evolution_time(
     # Start from a time whose signed [-pi, pi] phase interval contains the
     # spectrum bounded by lambda. QubitOperator.schatten_norm is the sum of
     # absolute Pauli coefficients denoted lambda in the chapter derivation.
-    bound_time = np.pi / qubit_hamiltonian.schatten_norm
+    coefficient_norm = float(qubit_hamiltonian.schatten_norm)
+    if not np.isfinite(coefficient_norm) or coefficient_norm <= 0.0:
+        raise ValueError(
+            "qubit Hamiltonian coefficient norm must be positive and finite"
+        )
+    if not np.isfinite(reference_active_energy_hartree):
+        raise ValueError("reference active energy must be finite")
+    if not np.isfinite(target_energy_error_hartree):
+        raise ValueError("target energy error must be finite")
+
+    bound_time = np.pi / coefficient_norm
     reference_phase = (bound_time * reference_active_energy_hartree / (2 * np.pi)) % 1.0
     # Quantize the continuous reference phase to the nearest finite-bit grid
     # point. Modulo handles the wrap from the final grid point back to zero.
     grid_size = 2**num_phase_bits
     grid_index = round(reference_phase * grid_size) % grid_size
+    if grid_index == 0:
+        raise ValueError(
+            "reference energy rounds to the zero-angle phase-grid point; "
+            "increase num_phase_bits or choose a different time-selection strategy"
+        )
     grid_phase = grid_index / grid_size
 
     # Convert the grid fraction to the signed angle convention used by the QDK
@@ -188,9 +204,14 @@ def choose_reference_guided_evolution_time(
 
     # Solve grid_angle = t * (E_reference + target_error) for t. The measured
     # grid energy therefore differs from the reference by the requested offset.
-    evolution_time = grid_angle / (
-        reference_active_energy_hartree + target_energy_error_hartree
-    )
+    target_active_energy = reference_active_energy_hartree + target_energy_error_hartree
+    if target_active_energy == 0.0:
+        raise ValueError("reference energy plus target error must be nonzero")
+    evolution_time = grid_angle / target_active_energy
+    if not np.isfinite(evolution_time) or evolution_time <= 0.0:
+        raise ValueError(
+            "reference-guided time selection did not produce a positive finite time"
+        )
     grid_energy = grid_angle / evolution_time
     aligned_reference_phase = (
         evolution_time * reference_active_energy_hartree / (2 * np.pi)
@@ -254,6 +275,7 @@ def prepare_iqpe_problem(
     *,
     num_phase_bits: int = DEFAULT_NUM_PHASE_BITS,
     shots_per_bit: int = DEFAULT_SHOTS_PER_BIT,
+    trial_determinants: int = 4,
 ) -> IqpeProblem:
     """Prepare all shared inputs and circuits before expensive simulation.
 
@@ -266,6 +288,7 @@ def prepare_iqpe_problem(
     Args:
         num_phase_bits: Number of bitwise IQPE iteration circuits.
         shots_per_bit: Simulator shots used by each bit's majority vote.
+        trial_determinants: Determinants retained in the selected trial state.
 
     Returns:
         A complete problem definition ready for ``run_complete_iqpe``.
@@ -273,7 +296,7 @@ def prepare_iqpe_problem(
     mapping = run_qubit_mapping_workflow()
     # Reuse the trial-state support selected by the preceding workflow rather
     # than synthesizing a different state inside the phase-estimation script.
-    trial_state = run_trial_state_workflow((4,)).trial_states[0]
+    trial_state = run_trial_state_workflow((trial_determinants,)).trial_states[0]
     evolution_time = choose_reference_guided_evolution_time(
         mapping.qubit_hamiltonian,
         mapping.mapped_active_energy,
@@ -465,6 +488,7 @@ def run_iqpe_workflow(
     first_seed: int = DEFAULT_FIRST_SEED,
     num_phase_bits: int = DEFAULT_NUM_PHASE_BITS,
     shots_per_bit: int = DEFAULT_SHOTS_PER_BIT,
+    trial_determinants: int = 4,
 ) -> IqpeWorkflowResult:
     """Prepare and execute repeated complete IQPE runs with progress output.
 
@@ -473,6 +497,7 @@ def run_iqpe_workflow(
         first_seed: Seed for the first run; later runs use consecutive seeds.
         num_phase_bits: Number of bitwise iteration circuits.
         shots_per_bit: Shots used for each phase-bit majority vote.
+        trial_determinants: Determinants retained in the selected trial state.
 
     Returns:
         Prepared problem, every complete run, bitstring frequencies, unique mode,
@@ -487,6 +512,7 @@ def run_iqpe_workflow(
     problem = prepare_iqpe_problem(
         num_phase_bits=num_phase_bits,
         shots_per_bit=shots_per_bit,
+        trial_determinants=trial_determinants,
     )
     print_iqpe_settings(
         problem,
